requirements.txt
django-allauth==0.38.0
django-bootstrap-form==3.4
django-bootstrap4==0.0.8
django-ckeditor==6.0.0
django-cleanup==5.2.0
django-cors-headers==2.4.0
django-crispy-forms==1.8.1
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==1.1.0
django-paypal==1.0.0
django-phonenumber-field==2.2.0
django-request==1.5.4pyth
django-storages==1.9.1
django-timezone-field==4.2.1
django-urlcrypt==0.1.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0